photo-enhance-service/pyproject.toml
[project]
name = "photo-enhance-service"
version = "0.1.0"
description = "Photo enhancement and anime conversion service using AI models"
readme = "README.md"
requires-python = ">=3.8,<3.14"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "opencv-python>=4.8.0,<5.0.0",
    "numpy>=1.24.0,<2.0.0",
    "Pillow>=10.0.0",
    "onnxruntime>=1.15.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "scipy>=1.9.0",
    "scikit-image>=0.20.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/photo_enhance_service"]
